Add dummy transfer cmd.
BAREOS has one more command that the changer cmd need to at least accept
and that is the transfer command. As this makes little sense on a
vchanger the command is accepted but not implemented.

Also did some small tweaks to the code for readability that were already
done when we implemented our own version of the listall command in a
previous version of vchanger.